\name{PB12matrix}
\alias{PB12matrix}
\docType{data}
\title{Plackett-Burman 12-run design matrix}
\description{ A binary matrix (-1 or +1) for a 12-run Plackett Burman design}
\usage{data("PB12matrix")}
\format{
PB12matrix has 12 rows and 11 columns labelled A through K.
}
\examples{
data(PB12matrix)
}
\keyword{datasets}

