\name{PB12response}
\alias{PB12response}
\docType{data}
\title{Response Vector for PB12 design}
\description{ A constructed response vector for a 12-run Plackett-Burman design}
\usage{data("PB12response")}
\format{
PB12response is a vector of length 12.
}
\details{
The values of "PB12response" were 
generated using the following active effects: B=7, D=5, H=11, I=4 and K=6. The 
remaining columns were all set to be inactive (effects equal 0). 
}

\examples{
data(PB12response)
}
\keyword{datasets}

